# api/openapi.yaml
openapi: 3.0.3
info:
  title: HeartFolio API
  description: |
    HeartFolio transforms emotional pain and healing processes into verifiable soft-skill assets.

    ## JSON:API

    All endpoints conform to the [JSON:API v1.0](https://jsonapi.org/) specification.

    **Content-Type:** `application/vnd.api+json`

    ### Response envelope

    Every response is wrapped in a top-level `data` object (single resource) or `data` array (collection).

    ```json
    { "data": { "type": "users", "id": "abc123", "attributes": { ... } } }
    ```

    ### Error envelope

    All errors return an `errors` array instead of `data`.

    ```json
    { "errors": [{ "status": "404", "title": "Not Found", "detail": "badge not found" }] }
    ```

    ### Relationships

    Resources that embed related data include a `relationships` key whose value is a full
    resource object (not just a linkage), making it a single-request pattern.

    ### Authentication

    Protected endpoints require `Authorization: Bearer <token>`.
    Obtain a token from `POST /api/v1/users/login` or `POST /api/v1/users/register`.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Reusable schemas ──────────────────────────────────────────────────────
  schemas:
...
    ErrorObject:
      type: object
      properties:
        status:
          type: string
          example: "404"
        title:
          type: string
          example: "Not Found"
        detail:
          type: string
          example: "badge not found"
    ErrorDocument:
      type: object
      required: [errors]
      properties:
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorObject"

    # ── Attribute bags ────────────────────────────────────────────────
...
    QuestAttributes:
      type: object
      properties:
        name:
          type: string
          example: "Solitude Strength"
        description:
          type: string
          example: "Practice being comfortable alone for 30 minutes"
        category:
          type: string
          example: "self-care"
        exp_reward:
          type: integer
          format: int64
          example: 50
        mood_tags:
          type: array
          items:
            type: string
          example: ["lonely", "anxious"]
        difficulty:
          type: string
          enum: [easy, medium, hard]
          example: "easy"
        estimated_duration:
          type: integer
          format: int32
          description: Duration in minutes
          example: 30
        badge_level:
          type: integer
          format: int32
          description: Badge level awarded on completion (0 = no badge)
          example: 1
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserQuestAttributes:
      type: object
      properties:
        status:
          type: string
          enum: [active, completed, abandoned]
          example: "active"
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        exp_earned:
          type: integer
          format: int64
          example: 50

    # ── Resource objects ───────────────────────────────────────────────
...
    QuestResource:
      type: object
      properties:
        type:
          type: string
          enum: [quests]
        id:
          type: string
          example: "665f1a2b3c4d5e6f7a8b9c0d"
        attributes:
          $ref: "#/components/schemas/QuestAttributes"
    UserQuestResource:
      type: object
      properties:
        type:
          type: string
          enum: [user-quests]
        id:
          type: string
          example: "665f1a2b3c4d5e6f7a8b9c0d"
        attributes:
          $ref: "#/components/schemas/UserQuestAttributes"
    UserQuestDetailResource:
      type: object
      properties:
        type:
          type: string
          enum: [user-quests]
        id:
          type: string
        attributes:
          $ref: "#/components/schemas/UserQuestAttributes"
        relationships:
          type: object
          properties:
            quest:
              type: object
              properties:
                data:
                  $ref: "#/components/schemas/QuestResource"
    MessageResource:
      type: object
      properties:
        type:
          type: string
          example: "badges"
        id:
          type: string
        attributes:
          type: object
          properties:
            message:
              type: string
              example: "operation successful"

    # ── Top-level documents ────────────────────────────────────────────
...
    QuestDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/QuestResource"
    QuestCollectionDocument:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/QuestResource"
    UserQuestDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/UserQuestResource"
    UserQuestDetailCollectionDocument:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/UserQuestDetailResource"
    MessageDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/MessageResource"

  # ── Reusable responses ─────────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/ErrorDocument"
          example:
            errors:
              - status: "401"
                title: "Unauthorized"
                detail: "missing authorization header"
    Forbidden:
      description: Insufficient permissions
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/ErrorDocument"
          example:
            errors:
              - status: "403"
                title: "Forbidden"
                detail: "admin access required"
    NotFound:
      description: Resource not found
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/ErrorDocument"
          example:
            errors:
              - status: "404"
                title: "Not Found"
                detail: "resource not found"
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/ErrorDocument"

# ── Paths ──────────────────────────────────────────────────────────────────────
paths:
...
  /api/v1/quests/{id}:
    get:
      tags: [Quests]
      summary: Get quest by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quest
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/QuestDocument"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/quests/mood/{mood}:
    get:
      tags: [Quests]
      summary: Adaptive quests by mood tag
      description: |
        Returns active quests whose `mood_tags` array contains the given mood.
        This powers the **Adaptive Quest** system — the frontend passes the user's
        current emotional state (e.g. `lonely`, `anxious`, `burnout`) and receives
        a tailored quest list.
      parameters:
        - name: mood
          in: path
          required: true
          schema:
            type: string
          example: lonely
      responses:
        "200":
          description: Quests matching the mood
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/QuestCollectionDocument"

  # ── Quests (authenticated) ─────────────────────────────────────────────────
  /api/v1/quests/{id}/start:
    post:
      tags: [Quests]
      summary: Start a quest
      description: |
        Assigns the quest to the authenticated user with `active` status.
        A user cannot start the same quest if it is already active.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Quest ID (from the quest library)
      responses:
        "201":
          description: UserQuest created
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/UserQuestDocument"
              example:
                data:
                  type: user-quests
                  id: "abc123"
                  attributes:
                    status: active
                    started_at: "2024-01-01T10:00:00Z"
                    completed_at: null
                    exp_earned: 0
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Quest already active for this user
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/ErrorDocument"
  /api/v1/quests/{userQuestID}/complete:
    post:
      tags: [Quests]
      summary: Complete a quest
      description: |
        Marks the UserQuest as `completed` and triggers the **Reward Loop**:
        1. Awards `exp_reward` EXP to the user
        2. Recalculates user level (100 EXP = 1 level)
        3. If the quest has a linked `badge_id`, awards or upgrades that badge

        The `id` here is the **UserQuest ID** (returned by the start endpoint),
        not the Quest library ID.
      security:
        - BearerAuth: []
      parameters:
        - name: userQuestID
          in: path
          required: true
          schema:
            type: string
          description: UserQuest ID (from start response)
      responses:
        "200":
          description: Quest completed, EXP awarded
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/UserQuestDocument"
              example:
                data:
                  type: user-quests
                  id: "abc123"
                  attributes:
                    status: completed
                    started_at: "2024-01-01T10:00:00Z"
                    completed_at: "2024-01-01T10:30:00Z"
                    exp_earned: 50
        "400":
          description: Quest is not in active status
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/ErrorDocument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/quests/{userQuestID}/abandon:
    post:
      tags: [Quests]
      summary: Abandon a quest
      description: Marks the UserQuest as `abandoned`. No EXP is awarded.
      security:
        - BearerAuth: []
      parameters:
        - name: userQuestID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quest abandoned
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/UserQuestDocument"
        "400":
          description: Quest is not in active status
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/ErrorDocument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/quests/me/quests:
    get:
      tags: [Quests]
      summary: Get my quests
      description: |
        Returns all UserQuests for the current user, enriched with full quest details
        in the `relationships.quest` key. Optionally filter by `status`.
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, completed, abandoned]
      responses:
        "200":
          description: User quest collection with embedded quest details
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/UserQuestDetailCollectionDocument"
              example:
                data:
                  - type: user-quests
                    id: "abc123"
                    attributes:
                      status: active
                      started_at: "2024-01-01T10:00:00Z"
                      completed_at: null
                      exp_earned: 0
                    relationships:
                      quest:
                        data:
                          type: quests
                          id: "def456"
                          attributes:
                            name: "Solitude Strength"
                            category: "self-care"
                            exp_reward: 50
                            mood_tags: [lonely]
                            difficulty: easy
                            estimated_duration: 30
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Admin: Badges ──────────────────────────────────────────────────────────
...
  /api/v1/admin/quests/:
    get:
      tags: ["Admin: Quests"]
      summary: List all quests (including inactive)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Quest collection
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/QuestCollectionDocument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      tags: ["Admin: Quests"]
      summary: Create a quest
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, exp_reward]
              properties:
                name:
                  type: string
                  example: "Solitude Strength"
                description:
                  type: string
                category:
                  type: string
                  example: "self-care"
                exp_reward:
                  type: integer
                  format: int64
                  minimum: 1
                  example: 50
                mood_tags:
                  type: array
                  items:
                    type: string
                  example: [lonely, anxious]
                difficulty:
                  type: string
                  enum: [easy, medium, hard]
                estimated_duration:
                  type: integer
                  format: int32
                  description: Duration in minutes
                badge_id:
                  type: string
                  description: Optional badge to award on completion
                badge_level:
                  type: integer
                  format: int32
                  description: Badge level to award on completion
      responses:
        "201":
          description: Quest created
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/QuestDocument"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/quests/{id}:
    get:
      tags: ["Admin: Quests"]
      summary: Get quest by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quest
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/QuestDocument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: ["Admin: Quests"]
      summary: Update a quest
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                category:
                  type: string
                exp_reward:
                  type: integer
                  format: int64
                mood_tags:
                  type: array
                  items:
                    type: string
                difficulty:
                  type: string
                  enum: [easy, medium, hard]
                estimated_duration:
                  type: integer
                  format: int32
                badge_id:
                  type: string
                badge_level:
                  type: integer
                  format: int32
                is_active:
                  type: boolean
      responses:
        "200":
          description: Quest updated
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/MessageDocument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: ["Admin: Quests"]
      summary: Delete a quest
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quest deleted
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/MessageDocument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"